requirements.txt
streamlit>=1.28.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python-headless>=4.5.0  # Headless version for cloud deployments
Pillow>=8.0.0
numpy>=1.19.0
matplotlib>=3.3.0
